server/requirements.txt
loguru
Flask
flask_cors
PyYAML
scikit-learn
torch
torchvision